🦄 test & refactor: add tests/compatibility.cpp and use `g++ -std=c++20 -Wall -Wextra -Wold-style-cast -pedantic -o .\build\compatibility.exe .\tests\compatibility.cpp; .\build\compatibility.exe` to test compatibility and eliminated all warnings

+++ sources/fraction.hpp
@@ -134,7 +134,7 @@ public:
     /// Convert the fraction to double type.
     constexpr operator double() const
     {
-        return (double)numerator_ / denominator_;
+        return double(numerator_) / double(denominator_);
     }
 
     /*
